Use ParserUtils instead of Parser for more OOP
